feat: random polish with tests, format, lint
