+++ apps/build-onchain-apps/src/onchain/utils/address.test.ts
@@ -1,0 +1,8 @@
+import { getSlicedAddress } from './address';
+
+describe('getSlicedAddress', () => {
+  it('should return a string of class names', () => {
+    const address = getSlicedAddress('0x1234567890123456789012345678901234567890');
+    expect(address).toEqual('0x123...7890');
+  });
+});
