+++ apps/build-onchain-apps/src/onchain/utils/balance.test.ts
@@ -1,0 +1,8 @@
+import { getAccountBalance } from './balance';
+
+describe('getAccountBalance', () => {
+  it('should return a string of class names', () => {
+    const accountBalance = getAccountBalance({ formatted: '1234567890', symbol: 'test' });
+    expect(accountBalance).toEqual('12345 test');
+  });
+});
