+++ apps/build-onchain-apps/src/pages/mint/components/NotConnected.tsx
@@ -1,0 +1,7 @@
+import { Text } from '@radix-ui/themes';
+
+function NotConnected() {
+  return <Text size="5">Please connect your wallet to continue.</Text>;
+}
+
+export default NotConnected;
