+++ apps/build-onchain-apps/src/utils/classNames.test.ts
@@ -1,0 +1,8 @@
+import { classNames } from './classNames';
+
+describe('classNames', () => {
+  it('should return a string of class names', () => {
+    const className = classNames('test', 'test2');
+    expect(className).toEqual('test test2');
+  });
+});
