+++ apps/build-onchain-apps/src/utils/timestamp.test.ts
@@ -4,7 +4,8 @@ describe('convertBigIntTimestampToDate', () => {
   it('should convert a big int timestamp to a date', () => {
     const timestamp = 1614105600000;
     convertBigIntTimestampToDate(timestamp as unknown as bigint);
-    // TODO Fix later
-    expect(true).toEqual(true);
+    expect(convertBigIntTimestampToDate(timestamp as unknown as bigint)).toEqual(
+      '2021-02-23T00:00:00.000Z',
+    );
   });
 });
